fix: attempt to fix pivot validation
