+++ app/Http/Api/Field/Pivot/Wiki/ArtistMember/ArtistMemberMemberIdField.php
@@ -4,21 +4,16 @@ declare(strict_types=1);
 
 namespace App\Http\Api\Field\Pivot\Wiki\ArtistMember;
 
-use App\Contracts\Http\Api\Field\CreatableField;
 use App\Contracts\Http\Api\Field\SelectableField;
 use App\Http\Api\Field\Field;
 use App\Http\Api\Query\Query;
 use App\Http\Api\Schema\Schema;
-use App\Models\Wiki\Artist;
 use App\Pivots\Wiki\ArtistMember;
-use Illuminate\Database\Eloquent\Builder;
-use Illuminate\Http\Request;
-use Illuminate\Validation\Rule;
 
 /**
  * Class ArtistMemberMemberIdField.
  */
-class ArtistMemberMemberIdField extends Field implements CreatableField, SelectableField
+class ArtistMemberMemberIdField extends Field implements SelectableField
 {
     /**
      * Create a new field instance.
@@ -41,25 +36,5 @@ class ArtistMemberMemberIdField extends Field implements CreatableField, Selecta
     {
         // Needed to match member relation.
         return true;
-    }
-
-    /**
-     * Set the creation validation rules for the field.
-     *
-     * @param  Request  $request
-     * @return array
-     */
-    public function getCreationRules(Request $request): array
-    {
-        $artistField = ArtistMember::ATTRIBUTE_ARTIST;
-
-        return [
-            'required',
-            'integer',
-            "different:{$artistField}",
-            Rule::exists(Artist::class, Artist::ATTRIBUTE_ID),
-            Rule::unique(ArtistMember::class, ArtistMember::ATTRIBUTE_MEMBER)
-                ->where(fn (Builder $query) => $query->where(ArtistMember::ATTRIBUTE_ARTIST, $request->get(ArtistMember::ATTRIBUTE_ARTIST))),
-        ];
     }
 }
